[BE] Add: 1:1 채팅 구현 완료

See merge request s07-webmobile1-sub2/S07P12E104!179

+++ backend/src/main/java/com/ssafy/honjaya/api/request/ChatReq.java
@@ -18,7 +18,7 @@ import lombok.ToString;
 @ApiModel(value = "ChatReq", description = "채팅 Req")
 public class ChatReq {
 	@ApiModelProperty(value = "채팅방 번호")
-	private int chatroomNo;
+	private long chatroomNo;
 	
 	@ApiModelProperty(value = "유저 번호")
 	private int userNo;
